requirements.txt
# Flusso RAG Demo - Requirements

# Core Dependencies
google-generativeai==0.8.5

# Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0
